requirements.txt
python-dotenv==0.9.1
requests==2.28.1
terminaltables==3.1.10